drivers/sht2x: fix printf format for native

+++ drivers/l3gxxxx/l3gxxxx.c
@@ -14,6 +14,8 @@
  * @{
  */
 
+#include <inttypes.h>
+#include <stdint.h>
 #include <string.h>
 #include <stdlib.h>
 
@@ -37,7 +39,7 @@
 
 #define DEBUG_DEV(m, d, ...) \
         DEBUG("[l3gxxxx] %s dev=%" PRIxPTR ": " m "\n", \
-              __func__, (unsigned int)d, ## __VA_ARGS__)
+              __func__, (uintptr_t)d, ## __VA_ARGS__)
 
 #else /* ENABLE_DEBUG */
 
@@ -1032,7 +1034,7 @@ static int _read(const l3gxxxx_t *dev, uint8_t reg, uint8_t *data, uint8_t len)
 
         if (ENABLE_DEBUG) {
             printf("[l3gxxxx] %s dev=%" PRIxPTR ": read following bytes: ",
-                   __func__,  (unsigned int)dev);
+                   __func__,  (uintptr_t)dev);
             for (uint8_t i = 0; i < len; i++) {
                 printf("%02x ", data[i]);
             }
@@ -1059,7 +1061,7 @@ static int _read(const l3gxxxx_t *dev, uint8_t reg, uint8_t *data, uint8_t len)
 
         if (ENABLE_DEBUG) {
             printf("[l3gxxxx] %s dev=%" PRIxPTR ": read following bytes: ",
-                   __func__,  (unsigned int)dev);
+                   __func__,  (uintptr_t)dev);
             for (uint8_t i = 0; i < len; i++) {
                 printf("%02x ", data[i]);
             }
@@ -1090,7 +1092,7 @@ static int _write(const l3gxxxx_t *dev,
 
         if (ENABLE_DEBUG) {
             printf("[l3gxxxx] %s dev=%" PRIxPTR ": write following bytes: ",
-                   __func__,  (unsigned int)dev);
+                   __func__,  (uintptr_t)dev);
             for (uint8_t i = 0; i < len; i++) {
                 printf("%02x ", data[i]);
             }
@@ -1112,7 +1114,7 @@ static int _write(const l3gxxxx_t *dev,
 
         if (ENABLE_DEBUG) {
             printf("[l3gxxxx] %s dev=%" PRIxPTR ": write following bytes: ",
-                   __func__,  (unsigned int)dev);
+                   __func__,  (uintptr_t)dev);
             for (uint8_t i = 0; i < len; i++) {
                 printf("%02x ", data[i]);
             }
